Start screen, death counter, score counter

+++ Assets/Scripts/PlayerController.cs
@@ -8,22 +8,21 @@ public class PlayerController : MonoBehaviour
 
     public float speed = 5;
     public float jumpForce = 10;
-    private bool isFalling = false; 
+
+    private bool isFalling = false;
+
     protected Rigidbody2D rb2D;
     protected SpriteRenderer sr;
-    private float boxColliderSizeX;
-    private float boxColliderSizeY;
-    private Vector2 startingCollider;
-    //public Sprite crouchSprite;
-    public Sprite defaultSprite;
+
+    private DeathCounter dc;
+    private bool destroying;
 
     private void Start()
     {
         rb2D = GetComponent<Rigidbody2D>();
         sr = GetComponent<SpriteRenderer>();
-        //boxColliderSizeX = gameObject.GetComponent<BoxCollider2D>().size.x;
-        //boxColliderSizeY = gameObject.GetComponent<BoxCollider2D>().size.y;
-        //startingCollider = new Vector2(boxColliderSizeX, boxColliderSizeY);
+        dc = GameObject.FindObjectOfType<DeathCounter>();
+        destroying = false;
     }
 
     private void Update()
@@ -32,25 +31,12 @@ public class PlayerController : MonoBehaviour
         {
             float movementHorizontal = 0;
             Vector2 vel = rb2D.velocity;
-            //if (gameObject.GetComponent<BoxCollider2D>().size != startingCollider && !Input.GetKey(KeyCode.DownArrow))
-            //{
-            //    gameObject.GetComponent<BoxCollider2D>().size = startingCollider;
-            //    sr.sprite = defaultSprite;
-            //}
 
-            /*if (Input.GetKey(KeyCode.DownArrow))
-            {
-                sr.sprite = crouchSprite;
-                if (gameObject.GetComponent<BoxCollider2D>().size.y >= boxColliderSizeY)
-                {
-                    gameObject.GetComponent<BoxCollider2D>().size -= new Vector2(0, (.63f * boxColliderSizeY));
-                }
-            }*/
             if (Input.GetKeyDown(KeyCode.Space) && (vel.y == 0f || isFalling == true))
             {
                 vel.y = jumpForce;
             }
-            
+
             if (Input.GetKey(KeyCode.LeftArrow))
             {
                 movementHorizontal = -1;
@@ -68,8 +54,9 @@ public class PlayerController : MonoBehaviour
 
     private void OnCollisionEnter2D(Collision2D collision)
     {
-        if (collision.gameObject.CompareTag("Lava"))
+        if (collision.gameObject.CompareTag("Lava") && !destroying)
         {
+            destroying = true;
             destroyPlayer();
         }
         if (collision.gameObject.CompareTag("Falling"))
@@ -87,7 +74,6 @@ public class PlayerController : MonoBehaviour
     }
 
 
-
     public void destroyPlayer()
     {
         StartCoroutine(RestartTheGameAfterSeconds(1));
@@ -99,6 +85,9 @@ public class PlayerController : MonoBehaviour
     {
         yield return new WaitForSeconds(seconds);
         SceneManager.LoadScene(SceneManager.GetActiveScene().name);
+        dc.IncreaseDeaths();
+        destroying = false;
+
     }
 }
